YAML file to set the expanse report's fields (#12)
Script field_value_parser.py reads the values set in a YAML file and matches them to the expanse report's fields. the filtering of None values must be made recursive to eliminate None values from lists and dictionaries contained in other containers. That will be done in a module currently located in branch grv96_exploration.